refactor(shadow-reading): split assessment_result_dialog.dart (closes #585)
